requirements.txt
-i https://pypi.org/simple
aioboto3==11.2.0
aiobotocore[boto3]==2.5.0 ; python_version >= '3.7'
aiohttp==3.8.4 ; python_version >= '3.6'
aioitertools==0.11.0 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
anyio==3.7.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
awscrt==0.16.9
bcrypt==4.0.1
beanie==1.19.2
boto3==1.26.76
botocore==1.29.76 ; python_version >= '3.7'
bunnet==1.1.0
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
coverage[toml]==7.2.7 ; python_version >= '3.7'
cryptography==41.0.1 ; python_version >= '3.6'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==2.0.0.post2
exceptiongroup==1.1.1 ; python_version < '3.11'
fastapi==0.97.0
frozenlist==1.3.3 ; python_version >= '3.7'
gunicorn==20.1.0
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.17.2 ; python_version >= '3.7'
httptools==0.5.0
httpx==0.24.1
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
lazy-model==0.0.5 ; python_version >= '3.7' and python_version < '4.0'
markdown-it-py==3.0.0 ; python_version >= '3.8'
mdurl==0.1.2 ; python_version >= '3.7'
motor==3.1.2
multidict==6.0.4 ; python_version >= '3.7'
mypy==1.3.0
mypy-extensions==1.0.0 ; python_version >= '3.5'
numpy==1.24.3
packaging==23.1 ; python_version >= '3.7'
passlib[bcrypt]==1.7.4
pluggy==1.0.0 ; python_version >= '3.6'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
pygments==2.15.1 ; python_version >= '3.7'
pymongo[srv]==4.3.3
pytest==7.3.2
pytest-cov==4.1.0
pytest-lazy-fixture==0.6.3
pytest-order==1.1.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyyaml==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
s3transfer==0.6.1 ; python_version >= '3.7'
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1 ; python_version < '3.11'
tqdm==4.65.0
typing-extensions==4.6.3 ; python_version >= '3.7'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yarl==1.9.2 ; python_version >= '3.7'
